Got Forge Client build working.
