+++ common/src/main/java/com/negativelight/empoweredpathways/Constants.java
@@ -1,0 +1,11 @@
+package com.negativelight.empoweredpathways;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class Constants {
+
+	public static final String MOD_ID = "empoweredpathways";
+	public static final String MOD_NAME = "Empowered Pathways";
+	public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
+}
