+++ common/src/main/java/com/negativelight/empoweredpathways/platform/Services.java
@@ -1,7 +1,7 @@
-package com.example.examplemod.platform;
+package com.negativelight.empoweredpathways.platform;
 
-import com.example.examplemod.Constants;
-import com.example.examplemod.platform.services.IPlatformHelper;
+import com.negativelight.empoweredpathways.Constants;
+import com.negativelight.empoweredpathways.platform.services.IPlatformHelper;
 
 import java.util.ServiceLoader;
