+++ fabric/src/main/java/com/negativelight/empoweredpathways/platform/FabricPlatformHelper.java
@@ -1,6 +1,6 @@
-package com.example.examplemod.platform;
+package com.negativelight.empoweredpathways.platform;
 
-import com.example.examplemod.platform.services.IPlatformHelper;
+import com.negativelight.empoweredpathways.platform.services.IPlatformHelper;
 import net.fabricmc.loader.api.FabricLoader;
 
 public class FabricPlatformHelper implements IPlatformHelper {
