+++ fabric/src/main/resources/fabric.mod.json
@@ -1,6 +1,6 @@
 {
     "schemaVersion": 1,
-    "id": "mod_id",
+    "id": "empoweredpathways",
     "version": "${version}",
     "name": "${mod_name}",
     "description": "This is an example description! Tell everyone what your mod is about!",
@@ -11,17 +11,17 @@
         "homepage": "https://fabricmc.net/",
         "sources": "https://github.com/FabricMC/fabric-example-mod"
     },
-    "license": "CC0-1.0",
-    "icon": "assets/modid/icon.png",
+    "license": "All Rights Reserved",
+    "icon": "",
     "environment": "*",
     "entrypoints": {
         "main": [
-            "com.example.examplemod.ExampleMod"
+          "com.negativelight.empoweredpathways.EmpoweredPathways"
         ]
     },
     "mixins": [
-        "examplemod.mixins.json",
-        "examplemod.fabric.mixins.json"
+        "empoweredpathways.mixins.json",
+        "empoweredpathways.fabric.mixins.json"
     ],
     "depends": {
         "fabricloader": ">=0.14",
