+++ forge/src/main/java/com/negativelight/empoweredpathways/mixin/MixinTitleScreen.java
@@ -1,6 +1,6 @@
-package com.example.examplemod.mixin;
+package com.negativelight.empoweredpathways.mixin;
 
-import com.example.examplemod.Constants;
+import com.negativelight.empoweredpathways.Constants;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.TitleScreen;
 import org.spongepowered.asm.mixin.Mixin;
@@ -8,13 +8,13 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(Minecraft.class)
-public class MixinMinecraft {
-    
-    @Inject(at = @At("TAIL"), method = "<init>")
+@Mixin(TitleScreen.class)
+public class MixinTitleScreen {
+
+    @Inject(at = @At("HEAD"), method = "init()V")
     private void init(CallbackInfo info) {
-        
-        Constants.LOG.info("This line is printed by an example mod common mixin!");
+
+        Constants.LOG.info("This line is printed by an example mod mixin from Forge!");
         Constants.LOG.info("MC Version: {}", Minecraft.getInstance().getVersionType());
     }
 }
