+++ forge/src/main/resources/META-INF/mods.toml
@@ -3,24 +3,24 @@ loaderVersion = "[46,)" #mandatory This is typically bumped every Minecraft vers
 license = "C0-1.0" # Review your options at https://choosealicense.com/.
 #issueTrackerURL="https://change.me.to.your.issue.tracker.example.invalid/" #optional
 [[mods]] #mandatory
-modId = "multiloader" #mandatory
+modId = "empoweredpathways" #mandatory
 version = "${file.jarVersion}" #mandatory
 displayName = "${mod_name}" #mandatory
 #updateJSONURL="https://change.me.example.invalid/updates.json" #optional (see https://mcforge.readthedocs.io/en/latest/gettingstarted/autoupdate/)
 #displayURL="https://change.me.to.your.mods.homepage.example.invalid/" #optional (displayed in the mod UI)
-logoFile = "multiloader.png" #optional (needs to be in the root of your mod jar (root of your 'resources' folder))
+#logoFile = "multiloader.png" #optional (needs to be in the root of your mod jar (root of your 'resources' folder))
 credits = "Thanks for this example mod goes to Java" #optional
 authors = "${mod_author}" #optional
 description = '''
 This is a long form description of the mod. You can write whatever you want here
 ''' #mandatory (Supports multiline text)
-[[dependencies.multiloader]] #optional
+[[dependencies.empoweredpathways]] #optional
 modId = "forge" #mandatory
 mandatory = true #mandatory
 versionRange = "[46,)" #mandatory
 ordering = "NONE" # The order that this dependency should load in relation to your mod, required to be either 'BEFORE' or 'AFTER' if the dependency is not mandatory
 side = "BOTH" # Side this dependency is applied on - 'BOTH', 'CLIENT' or 'SERVER'
-[[dependencies.multiloader]]
+[[dependencies.empoweredpathways]]
 modId = "minecraft"
 mandatory = true
 versionRange = "[1.20,1.21)"
