--- forge/src/main/resources/META-INF/services/com.example.examplemod.platform.services.IPlatformHelper
@@ -1,1 +1,0 @@
-com.example.examplemod.platform.ForgePlatformHelper
