+++ forge/src/main/resources/META-INF/services/com.negativelight.empoweredpathways.platform.services.IPlatformHelper
@@ -1,0 +1,1 @@
+com.negativelight.empoweredpathways.platform.ForgePlatformHelper
