+++ forge/src/main/resources/empoweredpathways.forge.mixins.json
@@ -1,13 +1,13 @@
 {
     "required": true,
     "minVersion": "0.8",
-    "package": "com.example.examplemod.mixin",
+    "package": "com.negativelight.empoweredpathways.mixin",
     "refmap": "${mod_id}.refmap.json",
     "compatibilityLevel": "JAVA_17",
     "mixins": [
     ],
     "client": [
-        "MixinMinecraft"
+        "MixinTitleScreen"
     ],
     "server": [
     ],
@@ -15,4 +15,3 @@
         "defaultRequire": 1
     }
 }
-
