+++ gradle.properties
@@ -1,6 +1,6 @@
 # Project
-version=1.0.0
-group=com.example.examplemod
+version=0.0.0
+group=com.negativelight.empoweredpathways
 
 # Common
 minecraft_version=1.20
@@ -13,9 +13,9 @@ fabric_version=0.83.0+1.20
 fabric_loader_version=0.14.21
 
 # Mod options
-mod_name=ExampleMod
-mod_author=Jared
-mod_id=examplemod
+mod_name=EmpoweredPathways
+mod_author=Negative_light
+mod_id=empoweredpathways
 
 # Gradle
 org.gradle.jvmargs=-Xmx3G
