+++ settings.gradle
@@ -21,7 +21,7 @@ plugins {
 }
 
 // This should match the folder name of the project, or else IDEA may complain (see https://youtrack.jetbrains.com/issue/IDEA-317606)
-rootProject.name = 'MultiLoader-Template'
+rootProject.name = 'Empowered-Pathways'
 include("common")
 include("fabric")
 include("forge")
